+++ app/Http/Controllers/Account/CompaniesController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers\Account;
 
 use App\Http\Controllers\Controller;
+use App\Models\AgentAssignment;
 use App\Models\Company;
+use App\Models\Forecast;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -69,6 +71,14 @@ class CompaniesController extends Controller
 
     public function delete(Company $company)
     {
+        if(AgentAssignment::where('company_id', $company->id)->count() > 0) {
+            return redirect(route('account.companies.index'))->with(['error' => 'Company has assigned to agents. Please delete them first.']);
+        }
+
+        if(Forecast::where('company_id', $company->id)->count() > 0) {
+            return redirect(route('account.companies.index'))->with(['error' => 'Company has forecasts. Please delete them first.']);
+        }
+
         $company->delete();
 
         return redirect(route('account.companies.index'))->with(['success' => 'Company has been deleted successfully.']);
